# Conflicts:
#	app/src/main/java/com/lion/finalprojectshoppingmallservice3team/Component/LikeLionFilterChip.kt

+++ app/src/main/java/com/lion/finalprojectshoppingmallservice3team/Component/LikeLionFilterChip.kt
@@ -1,5 +1,6 @@
 package com.lion.finalprojectshoppingmallservice3team.Component
 
+import android.R.attr.text
 import android.annotation.SuppressLint
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Row
@@ -8,21 +9,11 @@ import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.Clear
-import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.Done
-import androidx.compose.material3.FilterChip
-import androidx.compose.material3.FilterChipDefaults
 import androidx.compose.material3.Icon
 import androidx.compose.material3.IconButton
 import androidx.compose.material3.Surface
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
@@ -42,7 +33,7 @@ fun LikeLionFilterChip(
     chipModifier: Modifier,
     modifier: Modifier = Modifier,
     onChipClicked: (String, Boolean) -> Unit,
-    onDeleteButtonClicked: (String) -> Unit  = {}// 삭제 버튼 콜백 추가
+    onDeleteButtonClicked: ((String) -> Unit)? = null // 삭제 버튼 콜백 추가
 ) {
 
     Surface(
@@ -68,16 +59,17 @@ fun LikeLionFilterChip(
                 modifier = chipModifier
                     .clickable { onChipClicked(text, selected) }
             )
-
-            IconButton(
-                onClick = { onDeleteButtonClicked(text) },
-                modifier = Modifier.size(20.dp)
-            ) {
-                Icon(
-                    imageVector = Icons.Default.Clear, // 삭제 아이콘 설정
-                    contentDescription = "Delete Chip",
-                    tint = Color.Gray
-                )
+            if (onDeleteButtonClicked!=null) {
+                IconButton(
+                    onClick = { onDeleteButtonClicked(text) },
+                    modifier = Modifier.size(20.dp)
+                ) {
+                    Icon(
+                        imageVector = Icons.Default.Clear, // 삭제 아이콘 설정
+                        contentDescription = "Delete Chip",
+                        tint = Color.Gray
+                    )
+                }
             }
         }
     }
